requirements.txt
torch>=2.1
torchaudio>=2.1
pytorch-lightning>=2.3.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0.1
scipy>=1.11
